+++ Changelog.md
@@ -7,6 +7,13 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
+## 198.0.0 - UNRELEASED
+
+### Fixed
+
+-   Product links for nRF54L15 DK, nRF54H20 DK, nRF9131 EK, nRF9151 DK, nRF9161
+    DK, and Nordic Thingy:91 X
+
 ## 197.0.0 - 2025-02-05
 
 ### Added